+++ src/components/pages/ProductCartPage/index.jsx
@@ -1,14 +1,8 @@
 import React, { useEffect, useState } from "react";
 
 import { useStore } from "hooks/useStore";
-import {
-  deleteCartList,
-  getCartList,
-  updateCartCount,
-} from "reducers/cartList";
+import { getCartList } from "reducers/cartList";
 
-import Spinner from "components/common/Spinner";
-import ErrorPage from "components/pages/ErrorPage";
 import PaymentAmount from "./PaymentAmount";
 import ProductCartList from "./ProductCartList";
 import {
@@ -19,54 +13,9 @@ import {
 } from "./styled";
 
 function ProductCartPage() {
-  const {
-    data: cartList,
-    isLoading,
-    errorMessage,
-    dispatch,
-  } = useStore("cartList");
+  const { data: cartList, dispatch } = useStore("cartList");
 
-  console.log(cartList);
   const [checkList, setCheckList] = useState([]);
-
-  const handleChangeAllCheckbox = () => {
-    if (checkList.length === 0) {
-      setCheckList(cartList.map((cartItem) => cartItem.id));
-      return;
-    }
-    setCheckList([]);
-  };
-
-  const handleDeleteAllItem = () => {
-    checkList.forEach((carItemId) => {
-      dispatch(deleteCartList(carItemId));
-    });
-    setCheckList([]);
-  };
-
-  const handleClickIncreaseButton = (id) => () => {
-    if (!checkList.includes(id)) setCheckList((prev) => [...prev, id]);
-    dispatch(updateCartCount(id, "increase"));
-  };
-
-  const handleClickDecreaseButton = (id, count) => () => {
-    if (count <= 1) return;
-    if (!checkList.includes(id)) setCheckList((prev) => [...prev, id]);
-    dispatch(updateCartCount(id, "decrease"));
-  };
-
-  const handleClickDeleteItemButton = (id) => () => {
-    dispatch(deleteCartList(id));
-    setCheckList((prev) => prev.filter((cartItemId) => cartItemId !== id));
-  };
-
-  const handleChangeCheckbox = (id) => () => {
-    if (checkList.includes(id)) {
-      setCheckList((prev) => prev.filter((cartItemId) => cartItemId !== id));
-      return;
-    }
-    setCheckList((prev) => [...prev, id]);
-  };
 
   const [totalPrice, totalCount] = cartList.reduce(
     (acc, { id, price, count }) => {
@@ -78,23 +27,6 @@ function ProductCartPage() {
     },
     [0, 0]
   );
-
-  const renderListContent = () => {
-    if (isLoading) return <Spinner />;
-    if (errorMessage) return <ErrorPage>에러: ${errorMessage} </ErrorPage>;
-    return (
-      <ProductCartList
-        cartList={cartList}
-        checkList={checkList}
-        handleChangeAllCheckbox={handleChangeAllCheckbox}
-        handleDeleteAllItem={handleDeleteAllItem}
-        handleClickIncreaseButton={handleClickIncreaseButton}
-        handleClickDecreaseButton={handleClickDecreaseButton}
-        handleClickDeleteItemButton={handleClickDeleteItemButton}
-        handleChangeCheckbox={handleChangeCheckbox}
-      />
-    );
-  };
 
   useEffect(() => {
     dispatch(getCartList());
@@ -108,7 +40,9 @@ function ProductCartPage() {
   return (
     <CartPageContainer>
       <CartPageHeader>장바구니</CartPageHeader>
-      <CartPageList>{renderListContent()}</CartPageList>
+      <CartPageList>
+        <ProductCartList checkList={checkList} setCheckList={setCheckList} />
+      </CartPageList>
       <CartPagePayment>
         <PaymentAmount
           position="sticky"
